change DeleteDisk behavior

+++ cloud/disk_manager/internal/pkg/services/disks/create_empty_disk_task.go
@@ -128,6 +128,13 @@ func (t *createEmptyDiskTask) Cancel(
 		)
 	}
 
+	if len(diskMeta.ZoneID) == 0 {
+		// If diskMeta has no zoneID, the disk was not in the database before
+		// calling storage.DeleteDisk, so it has already been marked as deleted.
+		// Need to call neither storage.DiskDeleted nor client.Delete.
+		return nil
+	}
+
 	err = client.Delete(ctx, t.params.Disk.DiskId)
 	if err != nil {
 		return err
